+++ test/Request/ArraySerializerTest.php
@@ -7,7 +7,8 @@
 
 namespace ZendTest\Diactoros\Request;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
+use UnexpectedValueException;
 use Zend\Diactoros\Request;
 use Zend\Diactoros\Request\ArraySerializer;
 use Zend\Diactoros\Stream;
@@ -110,7 +111,7 @@ class ArraySerializerTest extends TestCase
             ],
         ];
 
-        $this->setExpectedException('UnexpectedValueException');
+        $this->expectException(UnexpectedValueException::class);
 
         ArraySerializer::fromArray($serializedRequest);
     }