+++ test/Response/JsonResponseTest.php
@@ -9,7 +9,8 @@
 
 namespace ZendTest\Diactoros\Response;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 use Zend\Diactoros\Response\JsonResponse;
 
 class JsonResponseTest extends TestCase
@@ -70,28 +71,26 @@ class JsonResponseTest extends TestCase
         $this->assertEquals('foo/json', $response->getHeaderLine('content-type'));
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     */
     public function testJsonErrorHandlingOfResources()
     {
         // Serializing something that is not serializable.
         $resource = fopen('php://memory', 'r');
+
+        $this->expectException(InvalidArgumentException::class);
+
         new JsonResponse($resource);
     }
 
     public function testJsonErrorHandlingOfBadEmbeddedData()
     {
-        if (version_compare(PHP_VERSION, '5.5', 'lt')) {
-            $this->markTestSkipped('Skipped as PHP versions prior to 5.5 are noisy about JSON errors');
-        }
-
         // Serializing something that is not serializable.
         $data = [
             'stream' => fopen('php://memory', 'r'),
         ];
 
-        $this->setExpectedException('InvalidArgumentException', 'Unable to encode');
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Unable to encode');
+
         new JsonResponse($data);
     }
 
@@ -130,5 +129,66 @@ class JsonResponseTest extends TestCase
 
         $actual = json_decode($response->getBody()->getContents(), true);
         $this->assertEquals($json, $actual);
+    }
+
+    public function testPayloadGetter()
+    {
+        $payload = ['test' => 'data'];
+        $response = new JsonResponse($payload);
+        $this->assertEquals($payload, $response->getPayload());
+    }
+
+    public function testWithPayload()
+    {
+        $response = new JsonResponse(['test' => 'data']);
+        $json = [ 'foo' => 'bar'];
+        $newResponse = $response->withPayload($json);
+        $this->assertNotSame($response, $newResponse);
+
+        $this->assertEquals($json, $newResponse->getPayload());
+        $decodedBody = json_decode($newResponse->getBody()->getContents(), true);
+        $this->assertEquals($json, $decodedBody);
+    }
+
+    public function testEncodingOptionsGetter()
+    {
+        $response = new JsonResponse([]);
+        $this->assertSame(JsonResponse::DEFAULT_JSON_FLAGS, $response->getEncodingOptions());
+    }
+
+    public function testWithEncodingOptions()
+    {
+        $response = new JsonResponse([ 'foo' => 'bar']);
+        $expected = <<<JSON
+{"foo":"bar"}
+JSON;
+
+        $this->assertSame($expected, $response->getBody()->getContents());
+
+        $newResponse = $response->withEncodingOptions(JSON_PRETTY_PRINT);
+
+        $this->assertNotSame($response, $newResponse);
+
+        $expected = <<<JSON
+{
+    "foo": "bar"
+}
+JSON;
+
+        $this->assertSame($expected, $newResponse->getBody()->getContents());
+    }
+
+    public function testModifyingThePayloadDoesntMutateResponseInstance()
+    {
+        $payload = new \stdClass();
+        $payload->foo = 'bar';
+
+        $response = new JsonResponse($payload);
+
+        $originalPayload = clone $payload;
+        $payload->bar = 'baz';
+
+        $this->assertEquals($originalPayload, $response->getPayload());
+        $this->assertNotEquals($payload, $response->getPayload());
     }
 }